refactor: extract lifecycle/artifact contract types into homeboy-lifecycle-contract
Severs the lifecycle -> artifact_contract -> observation dependency chain that
blocked moving FuzzConfig/FuzzWorkloadConfig (and any lifecycle-carrying type)
out of core.

The ArtifactContract, EvidenceContract, and all LifecycleContract types are
pure serde data. Their only coupling to core was three conversion methods on
ArtifactContract (from_record, to_artifact_ref, From<ArtifactRef>) that
reference observation::ArtifactRecord and artifact_ref::ArtifactRef.

Splits data from behavior:
- New leaf crate homeboy-lifecycle-contract holds the pure structs + their
  self-contained methods (from_value, target, normalize) and schema consts.
- The three coupling conversions become free functions in core
  (artifact_contract_from_record, artifact_contract_to_ref,
  artifact_contract_from_ref). Only one caller (artifact_manifest.rs) needed
  rewiring; the From impl had no .into() callers.
- Core re-exports the types (pub use ...lifecycle, artifact_contract types) so
  crate::lifecycle::* and crate::artifact_contract::* paths stay stable.

This unblocks a future cut moving FuzzConfig/FuzzWorkloadConfig (and the rest of
the lifecycle-carrying extension config) to homeboy-extension-contract.
