+++ crates/homeboy-lifecycle-contract/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "homeboy-lifecycle-contract"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+authors = ["Chris Huber <chubes@extrachill.com>"]
+description = "Pure serializable artifact/evidence/lifecycle contract types for homeboy"
+repository = "https://github.com/Extra-Chill/homeboy"
+publish = false
+
+[lib]
+name = "homeboy_lifecycle_contract"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
