+++ crates/homeboy-lifecycle-contract/src/lifecycle.rs
